# Performance improvement for isFileBlacklisted by using caching