This PR improves the performance of ``isFileBlacklisted`` by using caching.
``isFileBlacklisted`` is called from many places and used quite often, especially from scanner.php