@DeepDiver1975 2015-04-19
> This PR improves the performance of isFileBlacklisted by using caching.

Can we please get a proof for this - e.g. send a comparison graph from blackfire.io

THX