@scrutinizer-notifier 2015-04-19
The [inspection](https://scrutinizer-ci.com/g/owncloud/core/inspections/46259fd4-ffc4-4f02-b850-19d3abaefce9) completed: **3 new issues**