@LukasReschke 2015-04-20
# Without memcache (100'000 runs)
With this change: 3.266398191452 sec
Without this change: 1.0600299835205 sec

=> Getting 2.2 secs *slower*

# With APCu enabled (100'000 runs)
With this change: 2.5553448200226 sec
Without this change: 1.0700287818909 sec

=>  Getting 1.5 secs *slower*
<hr/>
It actually makes things quite slower because querying the container is a quite expensive operation.

Please notice that actual execution time in prod. environments is faster than this measures since I had XDebug enabled.