@LukasReschke 2015-04-20
Test script:

```php
	for ($x = 0; $x <= 100000; $x++) {
		\OC\Files\Filesystem::isFileBlacklisted('aLongRandomFileName123');
	}

	$time2 = microtime(true);
	echo "script execution time: ".($time2-$time1); //value in seconds
```